feat(j): migrate customer.subscription.* handlers to registry (J-01b)
Why:
The 1197-line app/api/stripe/webhook/route.ts mixed dispatch, idempotency,
email helpers, and 7 event-specific handlers in a single switch — adding
or testing any one event meant reading the whole file. J-01a scaffolded
the typed registry; this iteration migrates the three subscription-state
handlers (customer.subscription.{created,updated,deleted}) plus their
shared upsertSubscription helper out of the central switch into the
per-event-family layout.

Verified callers / context:
- Local definitions of escapeHtml, sendTransactionalEmail, emailWrapper,
  build{ProWelcome,CourseReceipt,ConsultationConfirmation}Email, and
  upsertSubscription in route.ts were grepped: every remaining caller
  outside the 3 migrated cases (checkout.session.completed branches at
  L441, L550, L564, L639, L695, L769, L828, L857, L942, L971) imports
  the helpers from `@/lib/stripe-webhook/lib/email` instead. The local
  escapeHtml previously missed apostrophes; the SSOT @/lib/html-escape
  is more complete (bug-for-bug preservation isn't worth keeping).
- handlers/index.ts gains 3 registerHandler() calls; the comment
  placeholders for J-01c (charge.refunded, invoice.{paid,payment_failed},
  checkout.session.completed) remain.
- 47 stripe webhook tests pass against the new registry path
  (`__tests__/api/stripe-webhook.test.ts` + `-idempotency.test.ts`).

Idempotency / safety claim:
- upsertSubscription's out-of-order protection (existing.updated_at >
  stripeEventTime → skip) is preserved verbatim. The handler returns
  { status: "done" } regardless so the dispatch loop marks the
  idempotency row complete and Stripe stops retrying.
- Welcome / admin notification emails remain fire-and-forget with
  identical .catch swallowing — Resend outage doesn't block Stripe ack.
- Header-injection defence (escape + strip CR/LF) on the admin email
  subject is preserved.
- Behaviour for `.created` (welcome email + admin notification on
  active/trialing) and `.updated`/`.deleted` (upsert only) is byte-for-
  byte identical to the legacy switch arms.

Rollback location:
Revert this commit; the legacy switch arms are not yet deleted (J-01e
will remove them once all handlers are migrated and tested), so even
without the registry the route still works via fall-through.

Refs: #218
Audit: docs/audits/codebase-health-2026-04-24.md §J
Queue: docs/audits/REMEDIATION_QUEUE.md J-01b
